back/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==4.2
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Markdown==3.6
msgpack==1.0.8
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
redis==5.0.7
service-identity==24.1.0
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
websocket-client==1.8.0
zope.interface==6.4.post2
